@@ -8,7 +8,7 @@ environment:
 dependencies:
   cupertino_icons: ^1.0.2
   get: ^4.6.5
-  supabase_flutter: 1.2.2
+  supabase_flutter: 2.3.4
   flutter:
     sdk: flutter
   flutter_dotenv: ^5.0.2
@@ -26,6 +26,18 @@ dependencies:
   geocoding: ^2.1.0
   http: ^0.13.3
   image_picker: ^0.8.7+4
+   # Nuevas dependencias para la funcionalidad QR
+  qr_flutter: ^4.1.0
+  share_plus: ^7.0.2
+  path_provider: ^2.1.1
+  mobile_scanner: ^3.5.0 # Para móvil
+  universal_html: ^2.2.1 # Para detección de plataforma
+  image_picker_web: ^3.0.1
+  kd_utils: ^0.1.0+7 
+  flutter_local_notifications: ^18.0.1
+  shared_preferences: ^2.0.15
+  flutter_launcher_icons: ^0.13.1
+
 
 dev_dependencies:
   flutter_lints: ^2.0.0
@@ -37,3 +49,8 @@ flutter:
     - assets/
     - .env
   uses-material-design: true
+
+flutter_icons:
+  android: true
+  ios: true
+  image_path: "assets/Logo_color.png"
