@@ -1,14 +1,14 @@
 import 'package:euexia/app/data/models/usuarios.dart';
 
 class Foto {
-  int idFoto;
+  int? idFoto; // Ahora es opcional
   String urlFoto;
   int idUsuario;
-  
+
   Usuario? usuario;
 
   Foto({
-    required this.idFoto,
+    this.idFoto, // No es obligatorio
     required this.urlFoto,
     required this.idUsuario,
   });
@@ -22,10 +22,15 @@ class Foto {
   }
 
   Map<String, dynamic> toJson() {
-    return {
-      'idfoto': idFoto,
+    final data = {
       'urlfoto': urlFoto,
       'idusuario': idUsuario,
     };
+
+    if (idFoto != null) {
+      data['idfoto'] = idFoto.toString();
+    }
+
+    return data;
   }
 }
