@@ -1,4 +1,3 @@
-import 'dart:io';
 import 'dart:typed_data';
 import 'package:get/get.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
@@ -9,8 +8,11 @@ class GalleryController extends GetxController {
   var images = <String>[].obs;
   var isLoading = false.obs;
 
-  final SupabaseClient client = Supabase.instance.client;
+  final SupabaseClient _client = Supabase.instance.client;
   final ImagePicker _picker = ImagePicker();
+
+  // Obtener el UID del usuario actual
+  String get userUid => _client.auth.currentUser?.id ?? '';
 
   @override
   void onInit() {
@@ -19,19 +21,40 @@ class GalleryController extends GetxController {
   }
 
   Future<void> fetchImages() async {
+    if (userUid.isEmpty) {
+      Get.snackbar('Error', 'User not authenticated', 
+          colorText: Colors.white,
+          backgroundColor: Colors.red);
+      return;
+    }
+
     isLoading.value = true;
     try {
-      final response = await client.storage.from('gallery').list();
+      // Listar archivos solo en la carpeta del usuario
+      final response = await _client.storage.from('gallery').list(
+        path: userUid,
+      );
 
       if (response.isNotEmpty) {
         images.value = response.map((file) {
-          return client.storage.from('gallery').getPublicUrl(file.name);
+          return _client.storage
+              .from('gallery')
+              .getPublicUrl('$userUid/${file.name}');
         }).toList();
       } else {
-        Get.snackbar('Error', 'No images found', colorText: Colors.white);
+        images.clear();
+        Get.snackbar('Info', 'No images found in your gallery',
+            colorText: Colors.white,
+            backgroundColor: Colors.blue);
       }
     } catch (e) {
-      Get.snackbar('Error', 'An error occurred: $e', colorText: Colors.white);
+      if (e.toString().contains('Not found')) {
+        images.clear();
+      } else {
+        Get.snackbar('Error', 'Failed to load images: ${e.toString()}',
+            colorText: Colors.white,
+            backgroundColor: Colors.red);
+      }
     } finally {
       isLoading.value = false;
     }
@@ -93,38 +116,79 @@ class GalleryController extends GetxController {
 
   Future<void> _handleImagePick(ImageSource source) async {
     try {
-      if (client.auth.currentUser == null) {
-        Get.snackbar('Error', 'You must be logged in to upload photos');
+      if (userUid.isEmpty) {
+        Get.snackbar('Error', 'You must be logged in to upload photos',
+            colorText: Colors.white,
+            backgroundColor: Colors.red);
         return;
       }
 
       final XFile? image = await _picker.pickImage(source: source);
       if (image == null) return;
 
-      final Uint8List bytes = await image.readAsBytes();
       isLoading.value = true;
+      final Uint8List bytes = await image.readAsBytes();
 
-      final String fileName = 'photo_${DateTime.now().millisecondsSinceEpoch}.jpg';
+      // Generar nombre único para el archivo
+      final timestamp = DateTime.now().millisecondsSinceEpoch;
+      final extension = image.path.split('.').last.toLowerCase();
+      final fileName = 'image_$timestamp.$extension';
+      final filePath = '$userUid/$fileName';
 
-      if (GetPlatform.isWeb) {
-        await client.storage.from('gallery').uploadBinary(
-          fileName,
-          bytes,
-          fileOptions: const FileOptions(cacheControl: '3600', upsert: false),
-        );
-      } else {
-        final File imageFile = File(image.path);
-        await client.storage.from('gallery').upload(
-          fileName,
-          imageFile,
-          fileOptions: const FileOptions(cacheControl: '3600', upsert: false),
-        );
+      // Subir la imagen a la carpeta del usuario
+      await _client.storage.from('gallery').uploadBinary(
+            filePath,
+            bytes,
+            fileOptions: FileOptions(
+              cacheControl: '3600',
+              upsert: false,
+              contentType: 'image/$extension',
+            ),
+          );
+
+      // Actualizar la lista de imágenes
+      await fetchImages();
+      
+      Get.snackbar('Success', 'Photo uploaded successfully!',
+          colorText: Colors.white,
+          backgroundColor: Colors.green);
+    } catch (e) {
+      Get.snackbar('Error', 'Failed to upload photo: ${e.toString()}',
+          colorText: Colors.white,
+          backgroundColor: Colors.red);
+    } finally {
+      isLoading.value = false;
+    }
+  }
+
+  Future<void> deleteImage(String imageUrl) async {
+    try {
+      if (userUid.isEmpty) {
+        Get.snackbar('Error', 'You must be logged in to delete photos',
+            colorText: Colors.white,
+            backgroundColor: Colors.red);
+        return;
       }
 
+      // Extraer el nombre del archivo de la URL
+      final uri = Uri.parse(imageUrl);
+      final segments = uri.pathSegments;
+      final fileName = segments.last;
+      final filePath = '$userUid/$fileName';
+
+      isLoading.value = true;
+      await _client.storage.from('gallery').remove([filePath]);
+
+      // Actualizar la lista de imágenes
       await fetchImages();
-      Get.snackbar('Success', 'Photo uploaded successfully!', colorText: Colors.white);
+      
+      Get.snackbar('Success', 'Photo deleted successfully!',
+          colorText: Colors.white,
+          backgroundColor: Colors.green);
     } catch (e) {
-      Get.snackbar('Error', 'Failed to upload photo: $e', colorText: Colors.white);
+      Get.snackbar('Error', 'Failed to delete photo: ${e.toString()}',
+          colorText: Colors.white,
+          backgroundColor: Colors.red);
     } finally {
       isLoading.value = false;
     }
