@@ -1,29 +1,50 @@
 import 'package:get/get.dart';
+import 'package:supabase_flutter/supabase_flutter.dart';
+import 'package:euexia/app/modules/home/controllers/home_controller.dart';
+import 'package:euexia/app/routes/app_pages.dart';
 
 class SplashController extends GetxController {
-  // Add your variables and methods here
+  final _isLoading = true.obs;
+  bool get isLoading => _isLoading.value;
 
   @override
   void onInit() {
     super.onInit();
-    // Initialize your variables or call your methods here
+    _initializeApp();
   }
 
-  @override
-  void onReady() {
-    super.onReady();
-    // Called after the widget is rendered on screen
+  Future<void> _initializeApp() async {
+    try {
+      // Simula la carga de datos o realiza cualquier inicialización necesaria
+      await Future.delayed(Duration(seconds: 2));
+
+      // Redirige según el estado de autenticación
+      final supaProvider = Supabase.instance.client;
+      if (supaProvider.auth.currentUser == null) {
+        Get.offAllNamed(Routes.LOGIN);
+      } else {
+        Get.offAllNamed(Routes.HOME);
+      }
+    } catch (e) {
+      Get.snackbar('Error', 'Ocurrió un problema al inicializar la aplicación: $e');
+    }
+    finally {
+      _isLoading.value = false;
+    }
+  }
+
+  void redirectToAppropriateScreen() {
+    final supaProvider = Supabase.instance.client;
+    if (supaProvider.auth.currentUser == null) {
+      Get.offAllNamed(Routes.LOGIN);
+    } else {
+      Get.offAllNamed(Routes.HOME);
+    }
   }
 
   @override
   void onClose() {
-    // Clean up resources here
+    // Limpiar recursos si es necesario
     super.onClose();
-  }
-
-  // Example method
-  void navigateToHome() {
-    // Logic to navigate to the home page
-    Get.offAllNamed('/home');
   }
 }
