@@ -1,5 +1,3 @@
-import 'package:euexia/app/data/models/categorias.dart';
-import 'package:euexia/app/data/models/fotos.dart';
 import 'package:euexia/app/services/service.dart';
 import 'package:get/get.dart';
 import 'package:euexia/app/data/models/consejos.dart';
@@ -24,7 +22,7 @@ class TipsController extends GetxController {
 
   Future<void> getTips() async {
     isLoading.value = true;
-    result = await _supabaseService.ejercicios_rutinas.getEjerciciosOfRutina(1);
+    result = await _supabaseService.consejos.getTips();
 
     if(result.success){
       tips.assignAll(result.data as Iterable<Consejo>); // Asignar correctamente los datos a la lista observable
