@@ -1,11 +1,13 @@
 // ignore_for_file: constant_identifier_names
 
+import 'package:euexia/app/modules/exercises/bindings/exercises_binding.dart';
+import 'package:euexia/app/modules/exercises/views/exercises_view.dart';
 import 'package:euexia/app/modules/gallery/bindings/gallery_binding.dart';
+import 'package:euexia/app/modules/trainings/binding/trainings_binding.dart';
+import 'package:euexia/app/modules/trainings/view/trainings_view.dart';
 import 'package:get/get.dart';
 import 'package:euexia/app/modules/splash/bindings/splash_binding.dart';
 import 'package:euexia/app/modules/splash/views/splash_view.dart';
-import 'package:euexia/app/modules/add_note/bindings/add_note_binding.dart';
-import 'package:euexia/app/modules/add_note/views/add_note_view.dart';
 import 'package:euexia/app/modules/home/bindings/home_binding.dart';
 import 'package:euexia/app/modules/home/views/home_view.dart';
 import 'package:euexia/app/modules/login/bindings/login_binding.dart';
@@ -19,7 +21,6 @@ import 'package:euexia/app/modules/forgot_password/bindings/forgot_password_bind
 import 'package:euexia/app/modules/reset_password/views/reset_password_view.dart';
 import 'package:euexia/app/modules/reset_password/bindings/reset_password_binding.dart';
 import 'package:euexia/app/modules/gallery/views/gallery_view.dart';
-import 'package:euexia/app/modules/gallery/bindings/gallery_binding.dart';
 import 'package:euexia/app/modules/map/bindings/map_bindings.dart';
 import 'package:euexia/app/modules/map/views/map_view.dart';
 import 'package:euexia/app/modules/tips/bindings/tips_binding.dart';
@@ -28,6 +29,8 @@ import 'package:euexia/app/modules/challenges/bindings/challenges_binding.dart';
 import 'package:euexia/app/modules/challenges/views/challenges_view.dart';
 import 'package:euexia/app/modules/profile/bindings/profile_binding.dart';
 import 'package:euexia/app/modules/profile/views/profile_view.dart';
+import 'package:euexia/app/modules/qr/bindings/qr_binding.dart';
+import 'package:euexia/app/modules/qr/views/qr_view.dart';
 /*
 import 'package:euexia/app/modules/exercises/bindings/exercises_binding.dart';
 import 'package:euexia/app/modules/exercises/views/exercises_view.dart';*/
@@ -76,11 +79,6 @@ class AppPages {
       binding: ResetPasswordBinding(),
     ),
     GetPage(
-      name: _Paths.ADD_NOTE,
-      page: () => AddNoteView(),
-      binding: AddNoteBinding(),
-    ),
-    GetPage(
       name: _Paths.GALLERY,
       page: () => GalleryView(),
       binding: GalleryBinding(),
@@ -104,7 +102,23 @@ class AppPages {
       name: _Paths.PROFILE,
       page: () => ProfileView(),
       binding: ProfileBinding(),
-    ), /*
+    ),
+    GetPage(
+      name: _Paths.RUTINAS,
+      page: () => TrainingsView(),
+      binding: TrainingsBinding(),
+    ),
+     GetPage(
+      name: _Paths.EJERCICIOS,
+      page: () => ExercisesView(),
+      binding: ExercisesBinding(),
+    ), 
+    GetPage(
+      name: _Paths.QR,
+      page: () => QrView(),
+      binding: QrBinding(),
+    )
+     /*
     GetPage(
       name: _Paths.EXERCISES,
       page: () => ExercisesView(),
