@@ -23,30 +23,9 @@ void main() async {
   runApp(
     GetMaterialApp(
       title: "Application",
-      initialRoute: Routes.HOME, // Cambia la ruta inicial al splash
+      initialRoute: Routes.SPLASH, // Cambia la ruta inicial al splash
       getPages: AppPages.routes,
       debugShowCheckedModeBanner: false,
     ),
   );
 }
-
-// lib/main.dart
-//import 'package:flutter/material.dart';
-//import 'package:get/get.dart';
-//import 'bindings/profile_binding.dart';
-//import 'views/profile_view.dart';
-
-//void main() {
-  //runApp(MyApp());
-//}
-
-//class MyApp extends StatelessWidget {
-  //@override
-  //Widget build(BuildContext context) {
-    //return GetMaterialApp(
-      //debugShowCheckedModeBanner: false,
-      //initialBinding: ProfileBinding(),
-      //home: ProfileView(),
-    //);
-  //}
-//}
